+++ src/classes/child-processor.ts
@@ -1,6 +1,5 @@
 import { promisify } from 'util';
 import { JobJson, ParentCommand, SandboxedJob } from '../interfaces';
-
 import { childSend } from '../utils';
 
 enum ChildStatus {
@@ -100,7 +99,7 @@ export class ChildProcessor {
 // https://stackoverflow.com/questions/18391212/is-it-not-possible-to-stringify-an-error-using-json-stringify
 if (!('toJSON' in Error.prototype)) {
   Object.defineProperty(Error.prototype, 'toJSON', {
-    value: function () {
+    value: function toJSONByBull() {
       const alt: any = {};
       const _this = this;