+++ src/classes/queue-events.ts
@@ -1,4 +1,9 @@
-import { QueueEventsOptions, RedisClient, StreamReadRaw } from '../interfaces';
+import {
+  IoredisListener,
+  QueueEventsOptions,
+  RedisClient,
+  StreamReadRaw,
+} from '../interfaces';
 import {
   array2obj,
   clientCommandMessageReg,
@@ -8,7 +13,7 @@ import {
 import { QueueBase } from './queue-base';
 import { RedisConnection } from './redis-connection';
 
-export interface QueueEventsListener {
+export interface QueueEventsListener extends IoredisListener {
   /**
    * Listen to 'active' event.
    *
@@ -215,6 +220,10 @@ export class QueueEvents extends QueueBase {
     return this;
   }
 
+  /**
+   * Manually starts running the event consumming loop. This shall be used if you do not
+   * use the default "autorun" option on the constructor.
+   */
   async run(): Promise<void> {
     if (!this.running) {
       try {
@@ -283,6 +292,11 @@ export class QueueEvents extends QueueBase {
     }
   }
 
+  /**
+   * Stops consuming events and close the underlying Redis connection if necessary.
+   *
+   * @returns
+   */
   close(): Promise<void> {
     if (!this.closing) {
       this.closing = this.disconnect();