+++ src/classes/queue-scheduler.ts
@@ -1,4 +1,5 @@
 import {
+  IoredisListener,
   QueueSchedulerOptions,
   RedisClient,
   StreamReadRaw,
@@ -10,10 +11,9 @@ import {
   QUEUE_SCHEDULER_SUFFIX,
 } from '../utils';
 import { QueueBase } from './queue-base';
-import { Scripts } from './scripts';
 import { RedisConnection } from './redis-connection';
 
-export interface QueueSchedulerListener {
+export interface QueueSchedulerListener extends IoredisListener {
   /**
    * Listen to 'error' event.
    *
@@ -120,6 +120,15 @@ export class QueueScheduler extends QueueBase {
     return this;
   }
 
+  /**
+   * Run the scheduler.
+   *
+   * This method will run the scheduler in the background,
+   * and will not return until the scheduler is closed.
+   *
+   * If the QueueScheduler was instantiated with autorun: false,
+   * then it is not needed to manually call this method.
+   */
   async run(): Promise<void> {
     if (!this.running) {
       try {
@@ -214,6 +223,12 @@ export class QueueScheduler extends QueueBase {
     }
   }
 
+  /**
+   * isRunning
+   *
+   * @returns true if the scheduler is running
+   *
+   */
   isRunning(): boolean {
     return this.running;
   }
@@ -251,7 +266,7 @@ export class QueueScheduler extends QueueBase {
   private async updateDelaySet(timestamp: number): Promise<[number, string]> {
     if (!this.closing) {
       const result = await this.checkConnectionError(() =>
-        Scripts.updateDelaySet(this, timestamp),
+        this.scripts.updateDelaySet(timestamp),
       );
 
       if (!result) {
@@ -265,7 +280,7 @@ export class QueueScheduler extends QueueBase {
 
   private async moveStalledJobsToWait() {
     if (!this.closing) {
-      const [failed, stalled] = await Scripts.moveStalledJobsToWait(this);
+      const [failed, stalled] = await this.scripts.moveStalledJobsToWait();
 
       failed.forEach((jobId: string) =>
         this.emit(
@@ -279,6 +294,10 @@ export class QueueScheduler extends QueueBase {
     }
   }
 
+  /**
+   * Close the scheduler.
+   *
+   */
   close(): Promise<void> {
     if (this.closing) {
       return this.closing;