+++ src/commands/removeJob-1.lua
@@ -3,8 +3,8 @@
     In order to be able to remove a job, it cannot be active.
 
     Input:
-      KEYS[1] jobId
-      ARGV[1]  jobId
+      KEYS[1] queue prefix
+      ARGV[1] jobId
 
     Events:
       'removed'
@@ -14,33 +14,9 @@ local rcall = redis.call
 
 -- Includes
 --- @include "includes/destructureJobKey"
+--- @include "includes/isLocked"
+--- @include "includes/removeJobFromAnyState"
 --- @include "includes/removeParentDependencyKey"
-
--- recursively check if there are no locks on the
--- jobs to be removed.
-local function isLocked( prefix, jobId)
-    local jobKey = prefix .. jobId;
-
-    -- Check if this job is locked
-    local lockKey = jobKey .. ':lock'
-    local lock = rcall("GET", lockKey)
-    if not lock then
-        local dependencies = rcall("SMEMBERS", jobKey .. ":dependencies")
-        if (#dependencies > 0) then
-            for i, childJobKey in ipairs(dependencies) do
-                -- We need to get the jobId for this job.
-                local childJobId = getJobIdFromKey(childJobKey)
-                local childJobPrefix = getJobKeyPrefix(childJobKey, childJobId)
-                local result = isLocked( childJobPrefix, childJobId )
-                if result then
-                    return true
-                end
-            end
-        end
-        return false
-    end
-    return true
-end
 
 local function removeJob( prefix, jobId, parentKey)
     local jobKey = prefix .. jobId;
@@ -71,14 +47,9 @@ local function removeJob( prefix, jobId, parentKey)
         end
     end
 
-    rcall("LREM", prefix .. "active", 0, jobId)
-    rcall("LREM", prefix .. "wait", 0, jobId)
-    rcall("ZREM", prefix .. "delayed", jobId)
-    rcall("LREM", prefix .. "paused", 0, jobId)
-    rcall("ZREM", prefix .. "completed", jobId)
-    rcall("ZREM", prefix .. "failed", jobId)
+    local prev = removeJobFromAnyState(prefix, jobId)
+    
     rcall("ZREM", prefix .. "priority", jobId)
-    rcall("ZREM", prefix .. "waiting-children", jobId)
     rcall("DEL", jobKey, jobKey .. ":logs", jobKey .. ":dependencies", jobKey .. ":processed")
 
     -- -- delete keys related to rate limiter
@@ -89,10 +60,10 @@ local function removeJob( prefix, jobId, parentKey)
         --     rcall("HDEL", limiterIndexTable, jobId)
     -- end
 
-    rcall("XADD", prefix .. "events", "*", "event", "removed", "jobId", jobId, "prev", "unknown");
+    rcall("XADD", prefix .. "events", "*", "event", "removed", "jobId", jobId, "prev", prev);
 end
 
-local prefix = getJobKeyPrefix(KEYS[1], ARGV[1])
+local prefix = KEYS[1]
 
 if not isLocked(prefix, ARGV[1]) then
     removeJob(prefix, ARGV[1])