+++ tests/test_getters.ts
@@ -70,6 +70,33 @@ describe('Jobs getters', function () {
       await queueScheduler.close();
       await queueScheduler2.close();
     });
+
+    describe('when sharing connection', () => {
+      it('gets all queueSchedulers for this queue', async function () {
+        const ioredisConnection = new IORedis({ maxRetriesPerRequest: null });
+        const queueScheduler = new QueueScheduler(queueName, {
+          connection: ioredisConnection,
+        });
+        await queueScheduler.waitUntilReady();
+        await delay(10);
+
+        const queueSchedulers = await queue.getQueueSchedulers();
+        expect(queueSchedulers).to.have.length(1);
+
+        const queueScheduler2 = new QueueScheduler(queueName, {
+          connection: ioredisConnection,
+        });
+        await queueScheduler2.waitUntilReady();
+        await delay(10);
+
+        const nextQueueSchedulers = await queue.getQueueSchedulers();
+        expect(nextQueueSchedulers).to.have.length(2);
+
+        await queueScheduler.close();
+        await queueScheduler2.close();
+        await ioredisConnection.quit();
+      });
+    });
   });
 
   describe('.getWorkers', () => {
@@ -113,6 +140,36 @@ describe('Jobs getters', function () {
       await worker.close();
       await worker2.close();
       await removeAllQueueData(new IORedis(), queueName2);
+    });
+
+    describe('when sharing connection', () => {
+      it('gets all workers for this queue only', async function () {
+        const ioredisConnection = new IORedis({ maxRetriesPerRequest: null });
+        const worker = new Worker(queueName, async () => {}, {
+          connection: ioredisConnection,
+        });
+        const queueScheduler = new QueueScheduler(queueName, { connection });
+        await worker.waitUntilReady();
+        await queueScheduler.waitUntilReady();
+        await delay(10);
+
+        const workers = await queue.getWorkers();
+        expect(workers).to.have.length(1);
+
+        const worker2 = new Worker(queueName, async () => {}, {
+          connection: ioredisConnection,
+        });
+        await worker2.waitUntilReady();
+        await delay(10);
+
+        const nextWorkers = await queue.getWorkers();
+        expect(nextWorkers).to.have.length(2);
+
+        await worker.close();
+        await worker2.close();
+        await queueScheduler.close();
+        await ioredisConnection.quit();
+      });
     });
   });